third_party/WebKit/LayoutTests/animations/animationworklet/animate-non-accelerated-property-expected.txt

CONSOLE MESSAGE: line 62: background-color for the first target is: rgb(0, 128, 128)
CONSOLE MESSAGE: line 64: box-shadow for the second target is: rgb(0, 128, 128) 4px 4px 25px 0px
